Improved linc commands

+++ src/commands/quicklinc.ts
@@ -3,43 +3,7 @@ import { Command } from '../command';
 import { exec } from 'child_process';
 import * as fs from 'fs';
 import path from 'path';
-import crypto from 'crypto';
-
-function randomIdentifier(): string {
-    return crypto.randomBytes(64).toString('base64url');
-}
-
-async function cleanOutput(output: string, filename: string): Promise<string> {
-    return new Promise<string>((resolve, reject) => {
-        output = output.replace(/\x1B.*?m/gi, '');
-        output = output.split(filename).join('source.linc');
-        resolve(output);
-    });
-}
-
-async function runLinc(interaction: ChatInputCommandInteraction<CacheType>, filename: string): Promise<void> {
-    return new Promise<void>((resolve, reject) => {
-        exec(`timeout 5s firejail --quiet --whitelist=${filename} lincenv ${filename}`, async (error, stdout, stderr) => {
-            stdout = await cleanOutput(stdout, filename);
-            stderr = await cleanOutput(stderr, filename);
-            const error_string = await cleanOutput(error?.toString() || "", filename);
-            if(error){
-                await interaction.editReply({content: `\`\`\`stderr\n${error_string}\`\`\``});
-                resolve();
-                return;
-            }
-        
-            if(stderr){
-                await interaction.editReply({content: `\`\`\`stderr\n${stderr}\`\`\``});
-                resolve();
-                return;
-            }
-        
-            await interaction.editReply({content: `\`\`\`stdout\n${stdout}\`\`\``});
-            resolve();
-        });
-    });
-}
+import { randomIdentifier, runLinc } from '../linc';
 
 const QuickLinc : Command = {
     data: new SlashCommandBuilder()
@@ -47,18 +11,28 @@ const QuickLinc : Command = {
 		.setDescription('Execute linc code.')
         .setContexts(InteractionContextType.PrivateChannel)
         .addStringOption(option =>
-            option.setName('inline').setDescription('Evaluate linc code inline.').setRequired(true).setAutocomplete(false)
+            option.setName('code').setDescription('Evaluate linc code inline.').setRequired(true).setAutocomplete(false)
+        )
+        .addStringOption(option =>
+            option.setName('args').setDescription('Specify command line arguments for the program to execute with').setRequired(false)
         ),
     async execute(interaction, client) {
-        const input = interaction.options.getString('inline', true);
+        const input = interaction.options.getString('code', true);
+        const args = interaction.options.getString('args', false) || "";
+        const pattern = /[^A-Za-z0-9\-+*/<>@_'"\s=]/gi;
+        console.log(input);
         
         if(/system|sys_|read/gi.test(input))
-            return await interaction.reply('For safety reasons, syscall/file/shell/read interactions are disabled in this session.');
+            return await interaction.reply('For security reasons, syscall/file/shell/read interactions are disabled in this session.');
+        
+        if(pattern.test(args))
+            return await interaction.reply('For security reasons, the only allowed argument characters are A-Za-z0-9+-*/<>@_');
+        
         await interaction.deferReply();
         const relative_filename = path.join('.', 'temp', `${randomIdentifier()}.linc`);
         const filename = path.resolve(relative_filename);
         await fs.writeFileSync(filename, input);
-        await runLinc(interaction, filename);
+        await runLinc(interaction, filename, args);
         await fs.unlinkSync(filename);
     }
 };
